dynamodb-streams-lambda-filter/requirements.txt
aws-cdk-lib==2.3.0
constructs>=10.0.0,<11.0.0
boto3
pytz
